requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing and Augmentation
Pillow>=9.0.0
opencv-python>=4.7.0
albumentations==2.0.5

# Numerical Processing
numpy>=1.24.0
scipy>=1.10.0

# Data Analysis and Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Progress Bars and Logging
tqdm>=4.65.0
tensorboard>=2.12.0

# Metrics and Evaluation
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Utilities
PyYAML>=6.0
pathlib>=1.0.1

# Optional but recommended
jupyter>=1.0.0
ipython>=8.0.0
notebook>=6.5.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.3.0 